Adds link to Google Maps (or other geo: uri provider); extracts consts
In order to satisfy the license requirements for use of the mini map in
a non-browser context, this provides a link to the geo: uri for the
location (which Google Maps is). Note: this uses the old draft of the
geo: uri which encodes the coordinate in lat,lon. The new draft is not
yet supported by Google Maps.